fix: Prevent payee from deleting group transaction

+++ src/components/GroupTracker/GroupHistoryItem.js
@@ -1,15 +1,20 @@
-import React from "react";
+import React, { useState } from "react";
 import {
   Box,
   Typography,
   ListItem,
   ListItemText,
   IconButton,
+  Stack,
+  Alert,
+  Collapse,
 } from "@mui/material";
 import DeleteIcon from "@mui/icons-material/Delete";
+import CloseIcon from "@mui/icons-material/Close";
 
 import { collection, deleteDoc, getDocs } from "firebase/firestore";
 import { db } from "../../firebase-config";
+import { useAuth } from "../../context/AuthProvider";
 
 export const GroupHistoryItem = ({
   itemName,
@@ -19,9 +24,15 @@ export const GroupHistoryItem = ({
   from,
   to,
 }) => {
+  const { currentUsername } = useAuth();
   function handleDelete() {
-    deleteTransactionFromUsers(id);
-    deleteTransactionFromGroup(id);
+    if (currentUsername === to) {
+      setError("Only payer can delete transaction.");
+    } else {
+      setError("");
+      deleteTransactionFromUsers(id);
+      deleteTransactionFromGroup(id);
+    }
   }
 
   async function deleteTransactionFromUsers(id) {
@@ -55,50 +66,76 @@ export const GroupHistoryItem = ({
     });
   }
 
+  const [error, setError] = useState("");
+
   return (
-    <Box
-      sx={{
-        display: "flex",
-        gap: "8px",
-        alignItems: "center",
-        width: "clamp(250px, 50%, 400px)",
-        "&:hover": { "& .deleteIcon": { display: "block" } },
-      }}
-    >
-      <ListItem
+    <Stack spacing={0}>
+      {error && (
+        <Collapse in={error ? true : false}>
+          <Alert
+            severity="error"
+            action={
+              <IconButton
+                aria-label="close"
+                color="inherit"
+                size="small"
+                onClick={() => {
+                  setError("");
+                }}
+              >
+                <CloseIcon fontSize="inherit" />
+              </IconButton>
+            }
+            sx={{ my: 2 }}
+          >
+            {error}
+          </Alert>
+        </Collapse>
+      )}
+      <Box
         sx={{
-          backgroundColor: "common.white",
-          boxShadow: "1px 4px 5px 0px #EDEDED",
-          my: "4px",
-          border: "2px solid lightblue",
-          borderRadius: "8px",
+          display: "flex",
+          gap: "8px",
+          alignItems: "center",
+          width: "clamp(250px, 50%, 400px)",
+          "&:hover": { "& .deleteIcon": { display: "block" } },
         }}
       >
-        <ListItemText>
-          <Box
-            sx={{
-              display: "flex",
-              justifyContent: "space-between",
-              width: "100%",
-            }}
-          >
-            <Box>
-              <Typography variant="body1">{itemName}</Typography>
-              <Typography variant="body2">
-                {from}→{to}
-              </Typography>
+        <ListItem
+          sx={{
+            backgroundColor: "common.white",
+            boxShadow: "1px 4px 5px 0px #EDEDED",
+            my: "4px",
+            border: "2px solid lightblue",
+            borderRadius: "8px",
+          }}
+        >
+          <ListItemText>
+            <Box
+              sx={{
+                display: "flex",
+                justifyContent: "space-between",
+                width: "100%",
+              }}
+            >
+              <Box>
+                <Typography variant="body1">{itemName}</Typography>
+                <Typography variant="body2">
+                  {from}→{to}
+                </Typography>
+              </Box>
+              <Typography variant="body1">₹{amount}</Typography>
             </Box>
-            <Typography variant="body1">₹{amount}</Typography>
-          </Box>
-        </ListItemText>
-      </ListItem>
-      <IconButton
-        className="deleteIcon"
-        sx={{ height: "40px", width: "40px", display: "none" }}
-        onClick={handleDelete}
-      >
-        <DeleteIcon />
-      </IconButton>
-    </Box>
+          </ListItemText>
+        </ListItem>
+        <IconButton
+          className="deleteIcon"
+          sx={{ height: "40px", width: "40px", display: "none" }}
+          onClick={handleDelete}
+        >
+          <DeleteIcon />
+        </IconButton>
+      </Box>
+    </Stack>
   );
 };
